Edits DetailVC segueToMap function and moves convertTimeToDate and getWeatherIcon functions to DarkSkyFile
