+++ ScoutMaster/ScoutMaster/Controllers/MainTab/Dashboard/DetailVC.swift
@@ -268,8 +268,10 @@ class DetailVC: UIViewController, UIScrollViewDelegate {
         guard let windowScene = UIApplication.shared.connectedScenes.first as? UIWindowScene,
             let sceneDelegate = windowScene.delegate as? SceneDelegate, let window = sceneDelegate.window
             else { return }
-        if let tabBarController = window.rootViewController as? MainTabBarViewController {
+        if let tabBarController = window.rootViewController as? MainTabBarViewController, let mapVC = tabBarController.viewControllers![1] as? MapVC {
             tabBarController.selectedIndex = 1
+            mapVC.forecastDetails = self.forecastDetails
+            mapVC.trail = self.trail
         }
         dismiss(animated: true, completion: nil)
     }
@@ -603,48 +605,5 @@ extension DetailVC: UICollectionViewDataSource, UICollectionViewDelegateFlowLayo
     func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
         return CGSize(width: view.frame.width / 3, height: collectionView.frame.height * 0.9)
     }
-    
-    private func convertTimeToDate(forecastType: ForecastType, time: Int) -> String {
-        let dateInput = Date(timeIntervalSince1970: TimeInterval(exactly: time) ?? 0)
-        let formatter = DateFormatter()
-        switch forecastType {
-            case .daily:
-                formatter.dateFormat = "E M/d"
-            case .hourly:
-                formatter.dateFormat = "E h a"
-        }
-        formatter.locale = .current
-        return formatter.string(from: dateInput)
-    }
-    
-    private func getWeatherIcon(named: String) -> UIImage {
-        switch named {
-            case "clear-day":
-                return UIImage(systemName: "sun.max.fill")!
-            case "clear-night":
-                return UIImage(systemName: "moon.fill")!
-            case "wind":
-                return UIImage(systemName: "wind")!
-            case "rain":
-                return UIImage(systemName: "cloud.rain.fill")!
-            case "sleet":
-                return UIImage(systemName: "cloud.sleet.fill")!
-            case "snow":
-                return UIImage(systemName: "cloud.snow.fill")!
-            case "fog":
-                return UIImage(systemName: "cloud.fog.fill")!
-            case "cloudy":
-                return UIImage(systemName: "cloud.fill")!
-            case "hail":
-                return UIImage(systemName: "cloud.hail.fill")!
-            case "thunderstorm":
-                return UIImage(systemName: "cloud.bolt.rain.fill")!
-            case "tornado":
-                return UIImage(systemName: "tornado")!
-            default:
-                return UIImage(systemName: "cloud.sun.fill")!
-        }
-    }
-    
     
 }
